refactor(client): replace withAuth HOC with layout-based route protection
This commit refactors the authentication mechanism for protected routes by replacing the withAuth
Higher-Order Component (HOC) with a dedicated ProtectedLayout.
This change streamlines route protection and aligns with modern Next.js App Router conventions.

The new ProtectedLayout component now encapsulates the logic for checking user authentication
status. If the user is not authenticated, it automatically redirects them to the login page,
while displaying a loader during the verification process.

Key improvements:
  - Simplifies the protection of multiple routes by wrapping them in a single layout component.
  - Removes the need to individually wrap each page component with the withAuth HOC, resulting
    in cleaner and more declarative page implementations.
  - Centralizes the authentication guard logic, making it easier to maintain and update.
