--- apps/client/src/shared/HOC/index.ts
@@ -1,1 +1,0 @@
-export * from "./with-auth";
